Update README.md
Added a comment about the corrected Settings file which fixes a colourcoding bug that we had in previous versions. (RDFS:Class and owl:Class colours were swapped twice, in the code and settings file)
Using the new settings file for previous versions of Graph2VR would swap the two colours.